requirements.txt
requests
beautifulsoup4
pydantic
openai
python-dotenv
